+++ app/src/test/resources/json/latest_movie.json
@@ -1,0 +1,81 @@
+{
+  "adult": false,
+  "backdrop_path": "/5UkzNSOK561c2QRy2Zr4AkADzLT.jpg",
+  "belongs_to_collection": null,
+  "budget": 0,
+  "genres": [
+    {
+      "id": 878,
+      "name": "Science Fiction"
+    },
+    {
+      "id": 53,
+      "name": "Thriller"
+    },
+    {
+      "id": 18,
+      "name": "Drama"
+    }
+  ],
+  "homepage": "http://screen.nsw.gov.au/project/2067",
+  "id": 528085,
+  "imdb_id": "tt1918734",
+  "original_language": "en",
+  "original_title": "2067",
+  "overview": "A lowly utility worker is called to the future by a mysterious radio signal, he must leave his dying wife to embark on a journey that will force him to face his deepest fears in an attempt to change the fabric of reality and save humankind from its greatest environmental crisis yet.",
+  "popularity": 2257.847,
+  "poster_path": "/7D430eqZj8y3oVkLFfsWXGRcpEG.jpg",
+  "production_companies": [
+    {
+      "id": 22369,
+      "logo_path": null,
+      "name": "Arcadia",
+      "origin_country": "AU"
+    },
+    {
+      "id": 105250,
+      "logo_path": null,
+      "name": "KOJO Entertainment",
+      "origin_country": "AU"
+    },
+    {
+      "id": 142069,
+      "logo_path": null,
+      "name": "Elevate Production Finance",
+      "origin_country": "AU"
+    },
+    {
+      "id": 107909,
+      "logo_path": null,
+      "name": "Futurism Studios",
+      "origin_country": "US"
+    },
+    {
+      "id": 142070,
+      "logo_path": null,
+      "name": "Rocketboy",
+      "origin_country": "AU"
+    }
+  ],
+  "production_countries": [
+    {
+      "iso_3166_1": "AU",
+      "name": "Australia"
+    }
+  ],
+  "release_date": "2020-10-01",
+  "revenue": 0,
+  "runtime": 114,
+  "spoken_languages": [
+    {
+      "iso_639_1": "en",
+      "name": "English"
+    }
+  ],
+  "status": "Released",
+  "tagline": "The fight for the future has begun.",
+  "title": "2067",
+  "video": false,
+  "vote_average": 5.7,
+  "vote_count": 173
+}
