+++ app/src/test/resources/json/latest_tv_show.json
@@ -1,0 +1,151 @@
+{
+  "backdrop_path": "/g63KmFgqkvXu6WKS23V56hqEidh.jpg",
+  "created_by": [
+    {
+      "id": 68844,
+      "credit_id": "5ab47d649251417b020149ed",
+      "name": "Hayden Schlossberg",
+      "gender": 2,
+      "profile_path": null
+    },
+    {
+      "id": 347335,
+      "credit_id": "5ab36c76c3a368615b004c6e",
+      "name": "Josh Heald",
+      "gender": 2,
+      "profile_path": null
+    },
+    {
+      "id": 1801553,
+      "credit_id": "5ab47d690e0a265f24017c48",
+      "name": "John Hurwitz",
+      "gender": 2,
+      "profile_path": null
+    }
+  ],
+  "episode_run_time": [
+    30
+  ],
+  "first_air_date": "2018-05-02",
+  "genres": [
+    {
+      "id": 10759,
+      "name": "Action & Adventure"
+    },
+    {
+      "id": 18,
+      "name": "Drama"
+    }
+  ],
+  "homepage": "https://www.netflix.com/title/81002370",
+  "id": 77169,
+  "in_production": true,
+  "languages": [
+    "en"
+  ],
+  "last_air_date": "2019-04-24",
+  "last_episode_to_air": {
+    "air_date": "2019-04-24",
+    "episode_number": 10,
+    "id": 1768464,
+    "name": "No Mercy",
+    "overview": "It’s the first day of school at West Valley High but tensions between the two dojos have not cooled with the end of summer. And the rivalry between Daniel and Johnny comes to a head.",
+    "production_code": "",
+    "season_number": 2,
+    "show_id": 77169,
+    "still_path": "/kVXBPI1smqILLKgvnXoR74M9vRe.jpg",
+    "vote_average": 8.167,
+    "vote_count": 6
+  },
+  "name": "Cobra Kai",
+  "next_episode_to_air": {
+    "air_date": "2021-01-08",
+    "episode_number": 1,
+    "id": 2451489,
+    "name": "Episode 1",
+    "overview": "",
+    "production_code": "",
+    "season_number": 3,
+    "show_id": 77169,
+    "still_path": null,
+    "vote_average": 0,
+    "vote_count": 0
+  },
+  "networks": [
+    {
+      "name": "Netflix",
+      "id": 213,
+      "logo_path": "/wwemzKWzjKYJFfCeiB57q3r4Bcm.png",
+      "origin_country": ""
+    },
+    {
+      "name": "YouTube Premium",
+      "id": 1436,
+      "logo_path": "/3p05CgodUb9gPayuliuhawNj1Wo.png",
+      "origin_country": "US"
+    }
+  ],
+  "number_of_episodes": 21,
+  "number_of_seasons": 3,
+  "origin_country": [
+    "US"
+  ],
+  "original_language": "en",
+  "original_name": "Cobra Kai",
+  "overview": "This Karate Kid sequel series picks up 30 years after the events of the 1984 All Valley Karate Tournament and finds Johnny Lawrence on the hunt for redemption by reopening the infamous Cobra Kai karate dojo. This reignites his old rivalry with the successful Daniel LaRusso, who has been working to maintain the balance in his life without mentor Mr. Miyagi.",
+  "popularity": 2104.099,
+  "poster_path": "/eTMMU2rKpZRbo9ERytyhwatwAjz.jpg",
+  "production_companies": [
+    {
+      "id": 101024,
+      "logo_path": null,
+      "name": "Hurwitz & Schlossberg Productions",
+      "origin_country": "US"
+    },
+    {
+      "id": 11073,
+      "logo_path": "/wHs44fktdoj6c378ZbSWfzKsM2Z.png",
+      "name": "Sony Pictures Television",
+      "origin_country": "US"
+    },
+    {
+      "id": 907,
+      "logo_path": "/ca5SWI5uvU985f8Kbb4xc8AmVWH.png",
+      "name": "Overbrook Entertainment",
+      "origin_country": "US"
+    }
+  ],
+  "seasons": [
+    {
+      "air_date": "2018-05-02",
+      "episode_count": 10,
+      "id": 99459,
+      "name": "Season 1",
+      "overview": "Decades after the tournament that changed their lives, the rivalry between Johnny and Daniel reignites.",
+      "poster_path": "/lLrKXnM3WPEtrPCd1aTHT6x3hn.jpg",
+      "season_number": 1
+    },
+    {
+      "air_date": "2019-04-24",
+      "episode_count": 10,
+      "id": 120052,
+      "name": "Season 2",
+      "overview": "Johnny continues building a new life, but a face from his past could disrupt his future. Meanwhile, Daniel opens a Miyagi-Do studio to rival Cobra Kai.",
+      "poster_path": "/77kyNXN6yCRjDt9eBMj96vLMx8W.jpg",
+      "season_number": 2
+    },
+    {
+      "air_date": "2021-01-08",
+      "episode_count": 1,
+      "id": 160283,
+      "name": "Season 3",
+      "overview": "",
+      "poster_path": null,
+      "season_number": 3
+    }
+  ],
+  "status": "Returning Series",
+  "type": "Scripted",
+  "vote_average": 8,
+  "vote_count": 992
+}
