+++ core/src/main/java/me/dicoding/bajp/reel/core/utils/TestFixtureHelper.kt
@@ -1,4 +1,4 @@
-package me.dicoding.bajp.reel.core
+package me.dicoding.bajp.reel.core.utils
 
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
@@ -7,8 +7,11 @@ import me.dicoding.bajp.reel.core.data.model.json.MovieJson
 import me.dicoding.bajp.reel.core.data.model.json.MovieListJson
 import me.dicoding.bajp.reel.core.data.model.json.TvShowJson
 import me.dicoding.bajp.reel.core.data.model.json.TvShowListJson
+import okio.buffer
+import okio.source
+import java.nio.charset.StandardCharsets
 
-object JsonHelper {
+object TestFixtureHelper {
   val jsonBuilder = Json {
     ignoreUnknownKeys = true
     isLenient = true
@@ -28,4 +31,10 @@ object JsonHelper {
 
   fun loadFavoritesData(jsonString: String): List<FavoriteJson> =
     jsonBuilder.decodeFromString(jsonString)
+
+  fun parseStringFromJsonResource(fileName: String): String {
+    val inputStream = javaClass.classLoader!!.getResourceAsStream("json/$fileName")
+    val source = inputStream.source().buffer()
+    return source.readString(StandardCharsets.UTF_8)
+  }
 }
